+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.7.0
+ - added `shingle` command to `lsh` CLI;
+ - added poor man's suggestion for number of hash functions based on average sie of shingle sets;
+ - bumped `github.com/zoomio/inout` to `0.6.0`.
+
 ## 0.6.0
  - changes to CLI: use raw text lines of provided sources instead of tags from given sources.